+++ CHANGELOG.md
@@ -4,4 +4,44 @@
 
 ## [Unreleased]
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+### Version 1.5
+
+1. Now you can directly open JsonParser window from logcat or console window. No need to copy and paste from multiple window.
+2. Tree structure updated with child number and object number for arrays.
+3. Copy to clipboard added.
+
+### Version 1.4
+
+1. Multiple tabs added
+2. Rate/ donate tab issue fixed
+3. Menu item added under Edit menu (ctrl shift alt J)
+
+### Version 1.3
+
+1. Bug fixes
+
+### Version 1.2
+
+Added 3 options
+
+- Pretty print
+- Raw
+- Tree
+
+### Version 1.2.1
+
+1. Error message if not a valid json
+2. Bug fixes
+
+### Version 1.1
+
+Support for all platforms.
+
+### Version 1.0
+
+Parse VALID JSON string only.
+
+
+
+